requirement.txt
pillow
pillow-heif
opencv-python
requests
ollama